requirements.txt
mido~=1.3.2
PyYAML~=6.0.1
pynput~=1.7.7
cx_Freeze~=7.1.0